+++ custom_components/amazon_order_status/options_flow.py
@@ -13,8 +13,12 @@ from homeassistant.helpers.selector import (
 import voluptuous as vol
 
 from .const import (
+    CONF_EXPOSE_CARRIER,
+    CONF_EXPOSE_DELIVERY_DETAILS,
     CONF_EXPOSE_ITEM_TITLE,
+    CONF_EXPOSE_ITEM_IMAGE,
     CONF_EXPOSE_ORDER_ID,
+    CONF_EXPOSE_PARSER_DEBUG,
     CONF_EXPOSE_TRACKING_URL,
     CONF_IMAP_FOLDER,
     CONF_INITIAL_SCAN_DAYS,
@@ -75,6 +79,22 @@ class AmazonOrderStatusOptionsFlow(config_entries.OptionsFlow):
                 vol.Required(
                     CONF_EXPOSE_TRACKING_URL,
                     default=options.get(CONF_EXPOSE_TRACKING_URL, True),
+                ): BooleanSelector(),
+                vol.Required(
+                    CONF_EXPOSE_DELIVERY_DETAILS,
+                    default=options.get(CONF_EXPOSE_DELIVERY_DETAILS, False),
+                ): BooleanSelector(),
+                vol.Required(
+                    CONF_EXPOSE_CARRIER,
+                    default=options.get(CONF_EXPOSE_CARRIER, False),
+                ): BooleanSelector(),
+                vol.Required(
+                    CONF_EXPOSE_ITEM_IMAGE,
+                    default=options.get(CONF_EXPOSE_ITEM_IMAGE, False),
+                ): BooleanSelector(),
+                vol.Required(
+                    CONF_EXPOSE_PARSER_DEBUG,
+                    default=options.get(CONF_EXPOSE_PARSER_DEBUG, False),
                 ): BooleanSelector(),
                 vol.Optional(
                     CONF_IMAP_FOLDER,